+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/integrations/emi/ComputerRecipeHandler.kt
@@ -5,6 +5,7 @@ import dev.emi.emi.api.recipe.EmiPlayerInventory
 import dev.emi.emi.api.recipe.EmiRecipe
 import dev.emi.emi.api.recipe.handler.EmiCraftContext
 import dev.emi.emi.api.recipe.handler.EmiRecipeHandler
+import net.minecraft.client.Minecraft
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen
 import site.siredvin.peripheralworks.networking.ClientNetworking
 import site.siredvin.peripheralworks.networking.MapBasedEventMessage
@@ -23,7 +24,16 @@ class ComputerRecipeHandler<T : AbstractComputerMenu> : EmiRecipeHandler<T> {
         p0: EmiRecipe,
         p1: EmiCraftContext<T>,
     ): Boolean {
-        ClientNetworking.sendToServer(MapBasedEventMessage(p1.screenHandler, "emi_recipe_paste", CommonEntrypoint.mapRecipe(p0)))
+        ClientNetworking.sendToServer(
+            MapBasedEventMessage(
+                p1.screenHandler,
+                "emi_recipe_paste",
+                CommonEntrypoint.mapRecipe(
+                    p0,
+                    Minecraft.getInstance().level!!.registryAccess(),
+                ),
+            ),
+        )
         return true
     }
 }