+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/integrations/emi/EmiRecipePeripheralPlugin.kt
@@ -4,6 +4,7 @@ import dan200.computercraft.api.lua.LuaFunction
 import dan200.computercraft.api.lua.MethodResult
 import dev.emi.emi.registry.EmiRecipes
 import net.minecraft.resources.ResourceLocation
+import site.siredvin.broccolium.modules.platform.PlatformToolkit
 import site.siredvin.tweakium.modules.peripheral.api.IPeripheralPlugin
 
 class EmiRecipePeripheralPlugin : IPeripheralPlugin {
@@ -11,6 +12,6 @@ class EmiRecipePeripheralPlugin : IPeripheralPlugin {
     fun getEMI(recipeID: String): MethodResult {
         val id = ResourceLocation.tryParse(recipeID) ?: return MethodResult.of(null, "Incorrect recipe id")
         val recipe = EmiRecipes.manager.getRecipe(id) ?: return MethodResult.of(null, "Incorrect recipe id")
-        return MethodResult.of(CommonEntrypoint.mapRecipe(recipe))
+        return MethodResult.of(CommonEntrypoint.mapRecipe(recipe, PlatformToolkit.get().minecraftServer!!.registryAccess()))
     }
 }